requirements.txt
anyio==3.6.2
fastapi==0.95.1
graphql-core==3.2.3
idna==3.4
pydantic==1.10.7
python-dateutil==2.8.2
python-multipart==0.0.6
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
strawberry-graphql==0.172.0
typing_extensions==4.5.0
uvicorn==0.21.1
SQLAlchemy==2.0.9
asyncpg==0.27.0
aiofiles==23.1.0
alembic==1.10.4
psycopg2==2.9.6
pyyaml==6.0.1